moved functions around for dpp rework
