+++ scripts/check_known_pulsars.py
@@ -21,6 +21,7 @@ import checks
 import sn_flux_est as snfe
 from mwa_pulsar_client import client
 import submit_to_database as std
+import pulsar_obs_helper as poh
 
 import logging
 logger = logging.getLogger(__name__)
@@ -267,46 +268,6 @@ def get_pointings_required(source_ra, source_dec, fwhm, search_radius):
     pointing_list_list = fpio.format_ra_dec(temp, ra_col = 0, dec_col = 1)
     return pointing_list_list
 
-def find_pulsars_power(obsid, powers=None, names_ra_dec=None):
-    """
-    Finds the beam power information for pulsars in a specific obsid
-
-    Parameters:
-    -----------
-    obsid: int
-        The observation ID
-    powers: list/tuple
-        OPTIONAL - A list of minimum beam powers to evaluate the pulsar coverage at. If none, will use [0.3, 0.1]. Default: None
-    names_ra_dec: list
-        OPTIONAL - A list of puslars and their RA and Dec values to evaluate (generated from fpio.get_source_alog).
-                   If none, will look for all pulsars. Default: None
-
-    Returns:
-    --------
-    pulsar_power_dict: dictionary
-        Contains keys - power
-            Contains key - obsid
-                Contains one list for each pulsar found in that power
-                    Each list is constructed as [jname, enter, exit, max_power]
-    meta_data: list
-        A list of the output of get_common_obs_metadata for the input obsid
-    """
-    if not powers:
-        powers = [0.3, 0.1]
-    elif not (isinstance(powers, list) or isinstance(powers, tuple)):
-        #try this if powers isn't iterable
-        powers=list(powers)
-
-    if names_ra_dec is None:
-        names_ra_dec = np.array(fpio.grab_source_alog(max_dm=250))
-
-    pulsar_power_dict = {}
-    for pwr in powers:
-        obs_data, meta_data = fpio.find_sources_in_obs([obsid], names_ra_dec, dt_input=100, min_power=pwr)
-        pulsar_power_dict[pwr] = obs_data
-
-    return pulsar_power_dict, meta_data
-
 def get_sources_in_fov(obsid, source_type, fwhm):
     """
     Find all sources of the input type in the observations field-of-view
@@ -387,7 +348,7 @@ def submit_folds(obsid, DI_dir, cal_obs, args, psrbeg, psrend,
 
     #Find all pulsars in beam at at least 0.3 of zenith normlaized power
     names_ra_dec = np.array(fpio.grab_source_alog(max_dm=250))
-    pow_dict, meta_data = find_pulsars_power(obsid, powers=[0.3, 0.1], names_ra_dec=names_ra_dec)
+    pow_dict, meta_data = poh.find_pulsars_power(obsid, powers=[0.3, 0.1], names_ra_dec=names_ra_dec)
     channels = meta_data[-1][-1]
     obs_psrs = pow_dict[0.3][obsid]
     psrs_list_03 = [x[0] for x in obs_psrs]
